Adapt mserve_np.q to run with the batchmark "xapi" test runner. (use new "launchQ" utility)
The problem is files need to be accessed from different project repositories.

When a program is designed to load components using relative paths based in its own root,
it may still be launched from outside using an absolute path, however the components will
not be found using the relative paths as designed.

This problem may occur in 2 places when trying to launch a typical mserve test.
1. When mserve_np.q is launched by the batchmark testrunner (test-svr.q not found by mserve_np.q).
2. When test-svr.q (in the testrunner directory) is launched by mserve_np.q (data direcory not found by test-svr.q)

A new utility "launchQ" was developed to address this problem, by launching each project from its root directory.
It has been integrated into mserve and batchmark but not yet into launcher.q

This launch utility accepts a command string of the form: ( env-settings ) q-file-path arguments.
Where the optional environment settings, if included, must be enclosed in parentheses, and the "q" command itself is omitted.

1. The command string is parsed to separate the enviornment settings and parent directory frecom the command itself.
2. This allows a separate "cd" to the parent directory before the launch, and for the absolute path used to invoke
   "q" (available as .z.X 0) to be inserted between the environment settings and the actual command.
3. The previous current directory is restored after launch.
4. A base path may be specified to resolve the parent directory against, to support ~/ ../ etc.
5. Additional env settings may be specified to suppliment any provided in the command string.

The launchQ utility is affected by the very problem it was invented to avoid.
Because mserve_np.q expects to load launchQ.q from the current directory,
you must have launchQ.q in your current directory when you launch it.
This is no problem when you run it (after cd) from mserve root.
It may also be accomodated by using a symlink, as was done in each of the mserve/examples.

When invoking mserve_np.q from a batchmark test runner,
The dependencies are located in one of the mserve "example" or "demo" directories.
This directory will need to contain symlinks to mserve_np.q and launchQ.q,
and mserve_np.q will be launched using an absolute path to this symlink.

It would be nicer to invoke mserve_np.q from the mserve root directory, but that will require
consistently accessing dependencies in a common subdirectory, whereas now they are scattered through the
various examples (future project).

However, when the servant is loaded by mserve_np.q the base path for loading dependencies should be
the parent directory of the servant q-file, which is likely a batchmark project directory.

This is addressed by an environment variable: $LAUNCHQ_BASE.
When mserve_np.q launches a servant it will use this as the "base" path for launchQ.
When the batchmark test runner invokes mserve_np.q it will set this variable to its own parent directory.