Fix Bug in tickdemo.q: endofday process does not complete successfully when feed.q connects automatically after waiting for tick.q
The actual problem is that the "initdate" message is not received at startup
so "currentdate" never gets set, so the hdbAready and hdbBready messages will always
be ignored (as they need to be at startup). The solution is to set "currentdate"
when the "endofday" message is received, if it is not already set.